refactor: split rendering layer out of server module
src/server/ held both the axum HTTP layer and the rendering layer
(askama templates, view-model builders, i18n, embedded static assets).
cook build depends on the latter, so the two could not be gated apart.

Move rendering to src/web/ and lift the sync client to src/sync/ (login
and logout need it without the server). src/server/ now holds only the
axum layer, which lets the next commit put it behind a feature.

Pure code movement: no behaviour change.
