+++ templates/menu.html
@@ -144,7 +144,7 @@
                     {% for line in section.lines %}
                     {% if line.len() == 1 %}
                         {% match line[0] %}
-                        {% when crate::server::templates::MenuSectionItem::Text with (text) %}
+                        {% when crate::web::templates::MenuSectionItem::Text with (text) %}
                             {% if text.trim().ends_with(":") %}
                             <!-- Meal type header -->
                             <h3 class="text-lg font-semibold text-gray-800 mt-2 mb-1">{{ text }}</h3>
@@ -158,7 +158,7 @@
                             <!-- Single non-text item -->
                             <div class="pl-6 flex items-baseline gap-x-2">
                                 {% match line[0] %}
-                                {% when crate::server::templates::MenuSectionItem::RecipeReference with { name, scale } %}
+                                {% when crate::web::templates::MenuSectionItem::RecipeReference with { name, scale } %}
                                     <span class="inline-flex items-center gap-1">
                                         <a href="{{ prefix }}/recipe/{{ name.strip_prefix("./").unwrap_or(name) }}{% if static_mode %}.html{% endif %}"
                                            class="text-purple-600 hover:text-purple-800 font-medium hover:underline">
@@ -170,7 +170,7 @@
                                         {% when None %}
                                         {% endmatch %}
                                     </span>
-                                {% when crate::server::templates::MenuSectionItem::Ingredient with { name, quantity, unit } %}
+                                {% when crate::web::templates::MenuSectionItem::Ingredient with { name, quantity, unit } %}
                                     <span class="inline-flex items-center gap-1">
                                         <span class="ingredient-badge">
                                             {{ name }}
@@ -195,10 +195,10 @@
                     <div class="pl-6 flex flex-wrap items-baseline gap-x-2">
                         {% for item in line %}
                             {% match item %}
-                            {% when crate::server::templates::MenuSectionItem::Text with (text) %}
+                            {% when crate::web::templates::MenuSectionItem::Text with (text) %}
                                 <span class="text-gray-700">{{ text }}</span>
 
-                            {% when crate::server::templates::MenuSectionItem::RecipeReference with { name, scale } %}
+                            {% when crate::web::templates::MenuSectionItem::RecipeReference with { name, scale } %}
                                 <span class="inline-flex items-center gap-1">
                                     <a href="{{ prefix }}/recipe/{{ name.strip_prefix("./").unwrap_or(name) }}{% if static_mode %}.html{% endif %}"
                                        class="text-purple-600 hover:text-purple-800 font-medium hover:underline">
@@ -211,7 +211,7 @@
                                     {% endmatch %}
                                 </span>
 
-                            {% when crate::server::templates::MenuSectionItem::Ingredient with { name, quantity, unit } %}
+                            {% when crate::web::templates::MenuSectionItem::Ingredient with { name, quantity, unit } %}
                                 <span class="inline-flex items-center gap-1">
                                     <span class="ingredient-badge">
                                         {{ name }}
