+++ templates/recipe.html
@@ -317,7 +317,7 @@
                     <ol class="space-y-4 {% if !loop.first %}mt-4{% endif %}">
                         {% for item in section.items %}
                             {% match item %}
-                                {% when crate::server::templates::RecipeSectionItem::Step with (step) %}
+                                {% when crate::web::templates::RecipeSectionItem::Step with (step) %}
                                 <li class="bg-gradient-to-r from-gray-50 to-orange-50 rounded-xl p-4">
                                     <div class="flex flex-col gap-4">
                                         {% match step.image_path %}
@@ -331,7 +331,7 @@
                                                 <div class="text-gray-700 mb-2 leading-8">
                                                 {% for step_item in step.items %}
                                                     {% match step_item %}
-                                                        {% when crate::server::templates::StepItem::Text with (text) %}{{ text }}{% when crate::server::templates::StepItem::Ingredient with { name, reference_path } %}{% match reference_path %}{% when Some with (path) %}<a href="{{ prefix }}/recipe/{{ path }}{% if static_mode %}.html{% endif %}" class="ingredient-badge hover:underline" title="View recipe: {{ name }}">{{ name }}</a>{% when None %}<span class="ingredient-badge">{{ name }}</span>{% endmatch %}{% when crate::server::templates::StepItem::Cookware with (name) %}<span class="cookware-badge">{{ name }}</span>{% when crate::server::templates::StepItem::Timer with (name) %}<span class="timer-badge">⏱️ {{ name }}</span>{% when crate::server::templates::StepItem::Quantity with (qty) %}<span class="font-bold text-orange-600">{{ qty }}</span>{% when crate::server::templates::StepItem::LineBreak %}<br>{% endmatch %}{% endfor %}
+                                                        {% when crate::web::templates::StepItem::Text with (text) %}{{ text }}{% when crate::web::templates::StepItem::Ingredient with { name, reference_path } %}{% match reference_path %}{% when Some with (path) %}<a href="{{ prefix }}/recipe/{{ path }}{% if static_mode %}.html{% endif %}" class="ingredient-badge hover:underline" title="View recipe: {{ name }}">{{ name }}</a>{% when None %}<span class="ingredient-badge">{{ name }}</span>{% endmatch %}{% when crate::web::templates::StepItem::Cookware with (name) %}<span class="cookware-badge">{{ name }}</span>{% when crate::web::templates::StepItem::Timer with (name) %}<span class="timer-badge">⏱️ {{ name }}</span>{% when crate::web::templates::StepItem::Quantity with (qty) %}<span class="font-bold text-orange-600">{{ qty }}</span>{% when crate::web::templates::StepItem::LineBreak %}<br>{% endmatch %}{% endfor %}
                                                 </div>
                                                 {% if step.ingredients.len() > 0 %}
                                                 <div class="text-sm text-gray-600 mt-2 pl-4 border-l-2 border-orange-300">
@@ -346,7 +346,7 @@
                                         </div>
                                     </div>
                                 </li>
-                                {% when crate::server::templates::RecipeSectionItem::Note with (note) %}
+                                {% when crate::web::templates::RecipeSectionItem::Note with (note) %}
                                 <li class="bg-gradient-to-r from-blue-50 to-purple-50 rounded-xl p-4 mb-3 border-l-4 border-blue-400">
                                     <div class="flex items-start gap-2">
                                         <span class="text-blue-600">📝</span>
